* commit '20e47b2f901e3b8f0bd9624bf5e48c13c1d9a5be':
  [EDELIVERY-4959] the alert parameters are ignored in the backend anyway, we don't want their content to interfere with the blacklist validation
  [EDELIVERY-4959]  entityId of Alerts changed from Integer to Long

+++ Domibus-MSH/src/main/java/eu/domibus/web/rest/ro/AlertFilterRequestRO.java
@@ -14,7 +14,7 @@ public class AlertFilterRequestRO implements Serializable {
     private String processed;
     private String alertType;
     private String alertStatus;
-    private Integer alertId;
+    private Long alertId;
     private String alertLevel;
     private String creationFrom;
     private String creationTo;
@@ -81,11 +81,11 @@ public class AlertFilterRequestRO implements Serializable {
         this.alertStatus = alertStatus;
     }
 
-    public Integer getAlertId() {
+    public Long getAlertId() {
         return alertId;
     }
 
-    public void setAlertId(Integer alertId) {
+    public void setAlertId(Long alertId) {
         this.alertId = alertId;
     }
